# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Belong Phone Bank - OpenAPI 3.0
  description: |-
    This is a Phone Bank Service based on the OpenAPI 3.0 specification.  

  termsOfService: http://swagger.io/terms/
  contact:
    email: rangasanju@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:8080/phone-bank
tags:
  - name: Customer
    description: Customer Information
  - name: PhoneNumber
    description: Phone numbers for a customer
paths:

  /customer/phone-numbers/{customerId}:
    get:
      tags:
        - Customer
      summary: Returns all phone numbers for a given customer id
      description: Returns all phone numbers for a given customer id
      operationId: getCustomerPhoneNumbersById
      parameters:
        - name: customerId
          in: path
          description: ID of the customer
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneNumberResponse'

        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /customer/{customerId}:
    get:
      tags:
        - Customer
      summary: Returns Customer details for a given customer id
      description: Returns Customer details for a given customer id
      operationId: getCustomerById
      parameters:
        - name: customerId
          in: path
          description: ID of the customer
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseDto'

        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /phone-numbers:
    get:
      tags:
        - PhoneNumber
      summary: Returns all phone numbers
      description: Returns all phone numbers
      operationId: getPhoneNumbers
      parameters:
        - name: pageNo
          in: query
          description: Page No
          schema:
            type: integer
            format: int64
        - name: pageSize
          in: query
          description: Page Size
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneNumberResponseDto'

  /phone-numbers/{phoneNumberId}:
    get:
      tags:
        - PhoneNumber
      summary: Returns all phone numbers
      description: Returns all phone numbers
      operationId: getPhoneNumberById
      parameters:
        - name: phoneNumberId
          in: path
          description: phone id to be activated / deactivated
          required: true
          schema:
            type: string

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneNumber'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /phone-numbers/activate/{phoneNumberId}/{status}:
    put:
      tags:
        - PhoneNumber
      summary: Activate / Deactivate a phone number
      description: Activate a phone number
      operationId: activateNumber
      parameters:
        - name: phoneNumberId
          in: path
          description: phone id to be activated / deactivated
          required: true
          schema:
            type: string
        - name: status
          in: path
          description: phone id to be activated / deactivated
          required: true
          schema:
            type: boolean

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  phoneNumber:
                    $ref: '#/components/schemas/PhoneNumber'

        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    Meta:
      type: object
      properties:
        totalRecords:
          type: integer
          format: int64
          example: 10
        totalPages:
          type: integer
          format: int64
          example: 10

    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 100000
        firstName:
          type: string
          example: Tom
        lastName:
          type: string
          example: Hanks
        phoneNumbers:
          type: array
          items:
            $ref: '#/components/schemas/PhoneNumber'
      xml:
        name: customer

    PhoneNumber:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        number:
          type: string
          example: 111-222-333
        type:
          type: string
          example: WORK
        active:
          type: boolean
      xml:
        name: phonenumber

    PhoneNumberResponse:
      type: object
      properties:
        phoneNumbers:
          type: array
          items:
            $ref: '#/components/schemas/PhoneNumber'

    PhoneNumberResponseDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PhoneNumberResponse'
        Meta:
          $ref: '#/components/schemas/Meta'
    CustomerResponseDto:
      type: object
      properties:
        data:
          properties:
            customer:
              $ref: '#/components/schemas/Customer'

    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          example: 400
        message:
          type: string
          example: Some Error Message